Clean up ANSI warnings
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1405 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/nethack.d/unixtty.c
@@ -395,8 +395,8 @@ register char *bufp;
 				/* finish print our string */
 				putstr(bufp);
 				bufp = obufp; /* reset it */
-				if(strlen(obufp) < BUFSIZ-1 &&
-				 strlen(obufp) < COLNO)
+				if(strlen(obufp) < (unsigned)BUFSIZ-1 &&
+				 strlen(obufp) < (unsigned)COLNO)
 					/* set bufp at the end of our
 					 * string
 					 */
